requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.4
Brotli==1.1.0
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
clipboard==0.0.4
colorama==0.4.6
cryptography==42.0.7
distro==1.9.0
Django==5.0.1
djangorestframework==3.15.1
docopt==0.6.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fake-useragent==1.5.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.26.0
hyperframe==6.0.1
idna==3.6
kaitaistruct==0.10
MouseInfo==0.1.3
numpy==1.26.3
openai==1.11.0
opencv-python==4.9.0.80
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.0
pillow==10.2.0
pipreqs==0.4.13
pyarrow==16.0.0
pyasn1==0.6.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydivert==2.1.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyOpenSSL==24.1.0
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytweening==1.0.7
pytz==2024.1
requests==2.31.0
selenium==4.17.2
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tqdm==4.66.1
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.4
undetected-chromedriver==3.5.5
urllib3==2.2.0
webdriver-manager==4.0.1
websockets==12.0
wsproto==1.2.0
yarg==0.1.9
zstandard==0.22.0
